ZO-5358: Move datamanager helper to testing so we can reuse it
